+++ Temperance.Data/Models/HistoricalData/HistoricalPriceModel.cs
@@ -6,10 +6,10 @@
         public string Symbol { get; set; }
         public DateTime Timestamp { get; set; }
         public string TimeInterval { get; set; }
-        public decimal OpenPrice { get; set; }
-        public decimal HighPrice { get; set; }
-        public decimal LowPrice { get; set; }
-        public decimal ClosePrice { get; set; }
+        public double OpenPrice { get; set; }
+        public double HighPrice { get; set; }
+        public double LowPrice { get; set; }
+        public double ClosePrice { get; set; }
         public long Volume { get; set; }
     }
 }
